reversed some naming to make merging and tests easier. only check once for aesni
