+++ lib/chiapos/src/aes.hpp
@@ -1,270 +1,499 @@
-// Copyright 2018 Chia Network Inc
-
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-
-//    http://www.apache.org/licenses/LICENSE-2.0
-
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// Some public domain code is taken from pycrypto:
-// https://github.com/dlitz/pycrypto/blob/master/src/AESNI.c
-//
-//  AESNI.c: AES using AES-NI instructions
-//
-// Written in 2013 by Sebastian Ramacher <sebastian@ramacher.at>
-
-#ifndef SRC_CPP_AES_HPP_
-#define SRC_CPP_AES_HPP_
-
-#include <string.h>    // for memcmp
-#include <wmmintrin.h>   // for intrinsics for AES-NI
-
-/**
- *  Encrypts a message of 128 bits with a 128 bit key, using
- *  10 rounds of AES128 (9 full rounds and one final round). Uses AES-NI
- *  assembly instructions.
- */
-#define DO_ENC_BLOCK_128(m, k)              \
-    do                                      \
-    {                                       \
-        m = _mm_xor_si128(m, k[0]);         \
-        m = _mm_aesenc_si128(m, k[1]);      \
-        m = _mm_aesenc_si128(m, k[2]);      \
-        m = _mm_aesenc_si128(m, k[3]);      \
-        m = _mm_aesenc_si128(m, k[4]);      \
-        m = _mm_aesenc_si128(m, k[5]);      \
-        m = _mm_aesenc_si128(m, k[6]);      \
-        m = _mm_aesenc_si128(m, k[7]);      \
-        m = _mm_aesenc_si128(m, k[8]);      \
-        m = _mm_aesenc_si128(m, k[9]);      \
-        m = _mm_aesenclast_si128(m, k[10]); \
-    } while (0)
-
-/**
- *  Encrypts a message of 128 bits with a 256 bit key, using
- *  13 rounds of AES256 (13 full rounds and one final round). Uses
- *  AES-NI assembly instructions.
- */
-#define DO_ENC_BLOCK_256(m, k) \
-    do {\
-        m = _mm_xor_si128(m, k[ 0]); \
-        m = _mm_aesenc_si128(m, k[ 1]); \
-        m = _mm_aesenc_si128(m, k[ 2]); \
-        m = _mm_aesenc_si128(m, k[ 3]); \
-        m = _mm_aesenc_si128(m, k[ 4]); \
-        m = _mm_aesenc_si128(m, k[ 5]); \
-        m = _mm_aesenc_si128(m, k[ 6]); \
-        m = _mm_aesenc_si128(m, k[ 7]); \
-        m = _mm_aesenc_si128(m, k[ 8]); \
-        m = _mm_aesenc_si128(m, k[ 9]); \
-        m = _mm_aesenc_si128(m, k[ 10]);\
-        m = _mm_aesenc_si128(m, k[ 11]);\
-        m = _mm_aesenc_si128(m, k[ 12]);\
-        m = _mm_aesenc_si128(m, k[ 13]);\
-        m = _mm_aesenclast_si128(m, k[ 14]);\
-    }while(0)
-
-/**
- * Encrypts a message of 128 bits with a 128 bit key, using
- * 2 full rounds of AES128. Uses AES-NI assembly instructions.
- */
-#define DO_ENC_BLOCK_2ROUND(m, k)      \
-    do                                 \
-    {                                  \
-        m = _mm_xor_si128(m, k[0]);    \
-        m = _mm_aesenc_si128(m, k[1]); \
-        m = _mm_aesenc_si128(m, k[2]); \
-    } while (0)
-/**
- * Decrypts a ciphertext of 128 bits with a 128 bit key, using
- * 10 rounds of AES128 (9 full rounds and one final round).
- * Uses AES-NI assembly instructions.
- */
-#define DO_DEC_BLOCK(m, k)                  \
-    do                                      \
-    {                                       \
-        m = _mm_xor_si128(m, k[10 + 0]);    \
-        m = _mm_aesdec_si128(m, k[10 + 1]); \
-        m = _mm_aesdec_si128(m, k[10 + 2]); \
-        m = _mm_aesdec_si128(m, k[10 + 3]); \
-        m = _mm_aesdec_si128(m, k[10 + 4]); \
-        m = _mm_aesdec_si128(m, k[10 + 5]); \
-        m = _mm_aesdec_si128(m, k[10 + 6]); \
-        m = _mm_aesdec_si128(m, k[10 + 7]); \
-        m = _mm_aesdec_si128(m, k[10 + 8]); \
-        m = _mm_aesdec_si128(m, k[10 + 9]); \
-        m = _mm_aesdeclast_si128(m, k[0]);  \
-    } while (0)
-
-/**
- * Decrypts a ciphertext of 128 bits with a 128 bit key, using
- * 2 full rounds of AES128. Uses AES-NI assembly instructions.
- * Will not work unless key schedule is modified.
- */
 /*
-#define DO_DEC_BLOCK_2ROUND(m, k)           \
-    do                                      \
-    {                                       \
-        m = _mm_xor_si128(m, k[2 + 0]);    \
-        m = _mm_aesdec_si128(m, k[2 + 1]); \
-        m = _mm_aesdec_si128(m, k[2 + 2]); \
-    } while (0)
+
+This is an implementation of the AES algorithm, specifically ECB, CTR and CBC mode.
+Block size can be chosen in aes.h - available choices are AES128, AES192, AES256.
+
+The implementation is verified against the test vectors in:
+  National Institute of Standards and Technology Special Publication 800-38A 2001 ED
+
+ECB-AES128
+----------
+
+  plain-text:
+    6bc1bee22e409f96e93d7e117393172a
+    ae2d8a571e03ac9c9eb76fac45af8e51
+    30c81c46a35ce411e5fbc1191a0a52ef
+    f69f2445df4f9b17ad2b417be66c3710
+
+  key:
+    2b7e151628aed2a6abf7158809cf4f3c
+
+  resulting cipher
+    3ad77bb40d7a3660a89ecaf32466ef97 
+    f5d3d58503b9699de785895a96fdbaaf 
+    43b1cd7f598ece23881b00e3ed030688 
+    7b0c785e27e8ad3f8223207104725dd4 
+
+
+NOTE:   String length must be evenly divisible by 16byte (str_len % 16 == 0)
+        You should pad the end of the string with zeros if this is not the case.
+        For AES192/256 the key size is proportionally larger.
+
 */
 
-static __m128i key_schedule[20];  // The expanded key
 
-static __m128i aes128_keyexpand(__m128i key) {
-    key = _mm_xor_si128(key, _mm_slli_si128(key, 4));
-    key = _mm_xor_si128(key, _mm_slli_si128(key, 4));
-    return _mm_xor_si128(key, _mm_slli_si128(key, 4));
+/*****************************************************************************/
+/* Includes:                                                                 */
+/*****************************************************************************/
+#include <stdint.h>
+#include <string.h> // CBC mode, for memset
+
+//#define DISABLE_AESNI
+
+#ifndef DISABLE_AESNI
+#include <cpuid.h>
+#include "aesni.hpp"
+
+bool bHasAES=false;
+bool bCheckedAES=false;
+#endif // DISABLE_AESNI
+
+/*****************************************************************************/
+/* Defines:                                                                  */
+/*****************************************************************************/
+// The number of columns comprising a state in AES. This is a constant in AES. Value=4
+#define Nb 4
+
+/*****************************************************************************/
+/* Private variables:                                                        */
+/*****************************************************************************/
+// state - array holding the intermediate results during decryption.
+typedef uint8_t state_t[4][4];
+
+// The lookup-tables are marked const so they can be placed in read-only storage instead of RAM
+// The numbers below can be computed dynamically trading ROM for RAM - 
+// This can be useful in (embedded) bootloader applications, where ROM is often limited.
+static const uint8_t sbox[256] = {
+  //0     1    2      3     4    5     6     7      8    9     A      B    C     D     E     F
+  0x63, 0x7c, 0x77, 0x7b, 0xf2, 0x6b, 0x6f, 0xc5, 0x30, 0x01, 0x67, 0x2b, 0xfe, 0xd7, 0xab, 0x76,
+  0xca, 0x82, 0xc9, 0x7d, 0xfa, 0x59, 0x47, 0xf0, 0xad, 0xd4, 0xa2, 0xaf, 0x9c, 0xa4, 0x72, 0xc0,
+  0xb7, 0xfd, 0x93, 0x26, 0x36, 0x3f, 0xf7, 0xcc, 0x34, 0xa5, 0xe5, 0xf1, 0x71, 0xd8, 0x31, 0x15,
+  0x04, 0xc7, 0x23, 0xc3, 0x18, 0x96, 0x05, 0x9a, 0x07, 0x12, 0x80, 0xe2, 0xeb, 0x27, 0xb2, 0x75,
+  0x09, 0x83, 0x2c, 0x1a, 0x1b, 0x6e, 0x5a, 0xa0, 0x52, 0x3b, 0xd6, 0xb3, 0x29, 0xe3, 0x2f, 0x84,
+  0x53, 0xd1, 0x00, 0xed, 0x20, 0xfc, 0xb1, 0x5b, 0x6a, 0xcb, 0xbe, 0x39, 0x4a, 0x4c, 0x58, 0xcf,
+  0xd0, 0xef, 0xaa, 0xfb, 0x43, 0x4d, 0x33, 0x85, 0x45, 0xf9, 0x02, 0x7f, 0x50, 0x3c, 0x9f, 0xa8,
+  0x51, 0xa3, 0x40, 0x8f, 0x92, 0x9d, 0x38, 0xf5, 0xbc, 0xb6, 0xda, 0x21, 0x10, 0xff, 0xf3, 0xd2,
+  0xcd, 0x0c, 0x13, 0xec, 0x5f, 0x97, 0x44, 0x17, 0xc4, 0xa7, 0x7e, 0x3d, 0x64, 0x5d, 0x19, 0x73,
+  0x60, 0x81, 0x4f, 0xdc, 0x22, 0x2a, 0x90, 0x88, 0x46, 0xee, 0xb8, 0x14, 0xde, 0x5e, 0x0b, 0xdb,
+  0xe0, 0x32, 0x3a, 0x0a, 0x49, 0x06, 0x24, 0x5c, 0xc2, 0xd3, 0xac, 0x62, 0x91, 0x95, 0xe4, 0x79,
+  0xe7, 0xc8, 0x37, 0x6d, 0x8d, 0xd5, 0x4e, 0xa9, 0x6c, 0x56, 0xf4, 0xea, 0x65, 0x7a, 0xae, 0x08,
+  0xba, 0x78, 0x25, 0x2e, 0x1c, 0xa6, 0xb4, 0xc6, 0xe8, 0xdd, 0x74, 0x1f, 0x4b, 0xbd, 0x8b, 0x8a,
+  0x70, 0x3e, 0xb5, 0x66, 0x48, 0x03, 0xf6, 0x0e, 0x61, 0x35, 0x57, 0xb9, 0x86, 0xc1, 0x1d, 0x9e,
+  0xe1, 0xf8, 0x98, 0x11, 0x69, 0xd9, 0x8e, 0x94, 0x9b, 0x1e, 0x87, 0xe9, 0xce, 0x55, 0x28, 0xdf,
+  0x8c, 0xa1, 0x89, 0x0d, 0xbf, 0xe6, 0x42, 0x68, 0x41, 0x99, 0x2d, 0x0f, 0xb0, 0x54, 0xbb, 0x16 };
+
+static const uint8_t rsbox[256] = {
+  0x52, 0x09, 0x6a, 0xd5, 0x30, 0x36, 0xa5, 0x38, 0xbf, 0x40, 0xa3, 0x9e, 0x81, 0xf3, 0xd7, 0xfb,
+  0x7c, 0xe3, 0x39, 0x82, 0x9b, 0x2f, 0xff, 0x87, 0x34, 0x8e, 0x43, 0x44, 0xc4, 0xde, 0xe9, 0xcb,
+  0x54, 0x7b, 0x94, 0x32, 0xa6, 0xc2, 0x23, 0x3d, 0xee, 0x4c, 0x95, 0x0b, 0x42, 0xfa, 0xc3, 0x4e,
+  0x08, 0x2e, 0xa1, 0x66, 0x28, 0xd9, 0x24, 0xb2, 0x76, 0x5b, 0xa2, 0x49, 0x6d, 0x8b, 0xd1, 0x25,
+  0x72, 0xf8, 0xf6, 0x64, 0x86, 0x68, 0x98, 0x16, 0xd4, 0xa4, 0x5c, 0xcc, 0x5d, 0x65, 0xb6, 0x92,
+  0x6c, 0x70, 0x48, 0x50, 0xfd, 0xed, 0xb9, 0xda, 0x5e, 0x15, 0x46, 0x57, 0xa7, 0x8d, 0x9d, 0x84,
+  0x90, 0xd8, 0xab, 0x00, 0x8c, 0xbc, 0xd3, 0x0a, 0xf7, 0xe4, 0x58, 0x05, 0xb8, 0xb3, 0x45, 0x06,
+  0xd0, 0x2c, 0x1e, 0x8f, 0xca, 0x3f, 0x0f, 0x02, 0xc1, 0xaf, 0xbd, 0x03, 0x01, 0x13, 0x8a, 0x6b,
+  0x3a, 0x91, 0x11, 0x41, 0x4f, 0x67, 0xdc, 0xea, 0x97, 0xf2, 0xcf, 0xce, 0xf0, 0xb4, 0xe6, 0x73,
+  0x96, 0xac, 0x74, 0x22, 0xe7, 0xad, 0x35, 0x85, 0xe2, 0xf9, 0x37, 0xe8, 0x1c, 0x75, 0xdf, 0x6e,
+  0x47, 0xf1, 0x1a, 0x71, 0x1d, 0x29, 0xc5, 0x89, 0x6f, 0xb7, 0x62, 0x0e, 0xaa, 0x18, 0xbe, 0x1b,
+  0xfc, 0x56, 0x3e, 0x4b, 0xc6, 0xd2, 0x79, 0x20, 0x9a, 0xdb, 0xc0, 0xfe, 0x78, 0xcd, 0x5a, 0xf4,
+  0x1f, 0xdd, 0xa8, 0x33, 0x88, 0x07, 0xc7, 0x31, 0xb1, 0x12, 0x10, 0x59, 0x27, 0x80, 0xec, 0x5f,
+  0x60, 0x51, 0x7f, 0xa9, 0x19, 0xb5, 0x4a, 0x0d, 0x2d, 0xe5, 0x7a, 0x9f, 0x93, 0xc9, 0x9c, 0xef,
+  0xa0, 0xe0, 0x3b, 0x4d, 0xae, 0x2a, 0xf5, 0xb0, 0xc8, 0xeb, 0xbb, 0x3c, 0x83, 0x53, 0x99, 0x61,
+  0x17, 0x2b, 0x04, 0x7e, 0xba, 0x77, 0xd6, 0x26, 0xe1, 0x69, 0x14, 0x63, 0x55, 0x21, 0x0c, 0x7d };
+
+// The round constant word array, Rcon[i], contains the values given by 
+// x to the power (i-1) being powers of x (x is denoted as {02}) in the field GF(2^8)
+static const uint8_t Rcon[11] = {
+  0x8d, 0x01, 0x02, 0x04, 0x08, 0x10, 0x20, 0x40, 0x80, 0x1b, 0x36 };
+
+/*
+ * Jordan Goulder points out in PR #12 (https://github.com/kokke/tiny-AES-C/pull/12),
+ * that you can remove most of the elements in the Rcon array, because they are unused.
+ *
+ * From Wikipedia's article on the Rijndael key schedule @ https://en.wikipedia.org/wiki/Rijndael_key_schedule#Rcon
+ * 
+ * "Only the first some of these constants are actually used – up to rcon[10] for AES-128 (as 11 round keys are needed), 
+ *  up to rcon[8] for AES-192, up to rcon[7] for AES-256. rcon[0] is not used in AES algorithm."
+ */
+
+/*****************************************************************************/
+/* Private functions:                                                        */
+/*****************************************************************************/
+
+#define getSBoxValue(num) (sbox[(num)])
+
+
+// This function adds the round key to state.
+// The round key is added to the state by an XOR function.
+static void AddRoundKey(uint8_t round, state_t* state, const uint8_t* RoundKey)
+{
+  uint8_t i,j;
+  for (i = 0; i < 4; ++i)
+  {
+    for (j = 0; j < 4; ++j)
+    {
+      (*state)[i][j] ^= RoundKey[(round * Nb * 4) + (i * Nb) + j];
+    }
+  }
 }
 
-#define KEYEXP128_H(K1, K2, I, S) _mm_xor_si128(aes128_keyexpand(K1), \
-        _mm_shuffle_epi32(_mm_aeskeygenassist_si128(K2, I), S))
+// The SubBytes Function Substitutes the values in the
+// state matrix with values in an S-box.
+static void SubBytes(state_t* state)
+{
+  uint8_t i, j;
+  for (i = 0; i < 4; ++i)
+  {
+    for (j = 0; j < 4; ++j)
+    {
+      (*state)[j][i] = getSBoxValue((*state)[j][i]);
+    }
+  }
+}
 
-#define KEYEXP128(K, I) KEYEXP128_H(K, K, I, 0xff)
-#define KEYEXP256(K1, K2, I)  KEYEXP128_H(K1, K2, I, 0xff)
-#define KEYEXP256_2(K1, K2) KEYEXP128_H(K1, K2, 0x00, 0xaa)
+// The ShiftRows() function shifts the rows in the state to the left.
+// Each row is shifted with different offset.
+// Offset = Row number. So the first row is not shifted.
+static void ShiftRows(state_t* state)
+{
+  uint8_t temp;
 
-// public API
+  // Rotate first row 1 columns to left  
+  temp           = (*state)[0][1];
+  (*state)[0][1] = (*state)[1][1];
+  (*state)[1][1] = (*state)[2][1];
+  (*state)[2][1] = (*state)[3][1];
+  (*state)[3][1] = temp;
+
+  // Rotate second row 2 columns to left  
+  temp           = (*state)[0][2];
+  (*state)[0][2] = (*state)[2][2];
+  (*state)[2][2] = temp;
+
+  temp           = (*state)[1][2];
+  (*state)[1][2] = (*state)[3][2];
+  (*state)[3][2] = temp;
+
+  // Rotate third row 3 columns to left
+  temp           = (*state)[0][3];
+  (*state)[0][3] = (*state)[3][3];
+  (*state)[3][3] = (*state)[2][3];
+  (*state)[2][3] = (*state)[1][3];
+  (*state)[1][3] = temp;
+}
+
+static uint8_t xtime(uint8_t x)
+{
+  return ((x<<1) ^ (((x>>7) & 1) * 0x1b));
+}
+
+// MixColumns function mixes the columns of the state matrix
+static void MixColumns(state_t* state)
+{
+  uint8_t i;
+  uint8_t Tmp, Tm, t;
+  for (i = 0; i < 4; ++i)
+  {  
+    t   = (*state)[i][0];
+    Tmp = (*state)[i][0] ^ (*state)[i][1] ^ (*state)[i][2] ^ (*state)[i][3] ;
+    Tm  = (*state)[i][0] ^ (*state)[i][1] ; Tm = xtime(Tm);  (*state)[i][0] ^= Tm ^ Tmp ;
+    Tm  = (*state)[i][1] ^ (*state)[i][2] ; Tm = xtime(Tm);  (*state)[i][1] ^= Tm ^ Tmp ;
+    Tm  = (*state)[i][2] ^ (*state)[i][3] ; Tm = xtime(Tm);  (*state)[i][2] ^= Tm ^ Tmp ;
+    Tm  = (*state)[i][3] ^ t ;              Tm = xtime(Tm);  (*state)[i][3] ^= Tm ^ Tmp ;
+  }
+}
+
+// prints string as hex
+void phex(char *pre, uint8_t* str, int len)
+{
+    printf("%s %d ",pre,len);
+    
+    unsigned char i;
+    for (i = 0; i < len; ++i)
+        printf("%.2x", str[i]);
+    printf("\n");
+}
+
+uint8_t RoundKey128[176]; // AES_KEYLEN 16   Key length in bytes
+uint8_t RoundKey256[240]; // AES_KEYLEN 32   Key length in bytes
+
+#define KEYNR256 14
+#define KEYNR128 10
+
+#define KEYNK256 8
+#define KEYNK128 4
+
+#define ENCRYPTNR256 14
+#define ENCRYPTNR128 3
+
+// This function produces Nb(Nr+1) round keys. The round keys are used in each round to decrypt the states.
+static void KeyExpansion(uint8_t* RoundKey, const uint8_t* Key, int keyNr, int keyNk)
+{
+  unsigned i, j, k;
+  uint8_t tempa[4]; // Used for the column/row operations
+  
+  // The first round key is the key itself.
+  for (i = 0; i < keyNk; ++i)
+  {
+    RoundKey[(i * 4) + 0] = Key[(i * 4) + 0];
+    RoundKey[(i * 4) + 1] = Key[(i * 4) + 1];
+    RoundKey[(i * 4) + 2] = Key[(i * 4) + 2];
+    RoundKey[(i * 4) + 3] = Key[(i * 4) + 3];
+  }
+
+  // All other round keys are found from the previous round keys.
+  for (i = keyNk; i < Nb * (keyNr + 1); ++i)
+  {
+    {
+      k = (i - 1) * 4;
+      tempa[0]=RoundKey[k + 0];
+      tempa[1]=RoundKey[k + 1];
+      tempa[2]=RoundKey[k + 2];
+      tempa[3]=RoundKey[k + 3];
+
+    }
+
+    if (i % keyNk == 0)
+    {
+      // This function shifts the 4 bytes in a word to the left once.
+      // [a0,a1,a2,a3] becomes [a1,a2,a3,a0]
+
+      // Function RotWord()
+      {
+        const uint8_t u8tmp = tempa[0];
+        tempa[0] = tempa[1];
+        tempa[1] = tempa[2];
+        tempa[2] = tempa[3];
+        tempa[3] = u8tmp;
+      }
+
+      // SubWord() is a function that takes a four-byte input word and
+      // applies the S-box to each of the four bytes to produce an output word.
+
+      // Function Subword()
+      {
+        tempa[0] = getSBoxValue(tempa[0]);
+        tempa[1] = getSBoxValue(tempa[1]);
+        tempa[2] = getSBoxValue(tempa[2]);
+        tempa[3] = getSBoxValue(tempa[3]);
+      }
+
+      tempa[0] = tempa[0] ^ Rcon[i/keyNk];
+    }
+
+    // AES256 only
+    if ((keyNk==8)&&(i % keyNk == 4))
+    {
+      // Function Subword()
+      {
+        tempa[0] = getSBoxValue(tempa[0]);
+        tempa[1] = getSBoxValue(tempa[1]);
+        tempa[2] = getSBoxValue(tempa[2]);
+        tempa[3] = getSBoxValue(tempa[3]);
+      }
+    }
+
+    j = i * 4; k=(i - keyNk) * 4;
+    RoundKey[j + 0] = RoundKey[k + 0] ^ tempa[0];
+    RoundKey[j + 1] = RoundKey[k + 1] ^ tempa[1];
+    RoundKey[j + 2] = RoundKey[k + 2] ^ tempa[2];
+    RoundKey[j + 3] = RoundKey[k + 3] ^ tempa[3];
+  }
+}
 
 /*
  * Loads an AES key. Can either be a 16 byte or 32 byte bytearray.
  */
 void aes_load_key(uint8_t *enc_key, int keylen) {
-    switch (keylen) {
-        case 16: {
-            /* 128 bit key setup */
-            key_schedule[0] = _mm_loadu_si128((const __m128i*) enc_key);
-            key_schedule[1] = KEYEXP128(key_schedule[0], 0x01);
-            key_schedule[2] = KEYEXP128(key_schedule[1], 0x02);
-            key_schedule[3] = KEYEXP128(key_schedule[2], 0x04);
-            key_schedule[4] = KEYEXP128(key_schedule[3], 0x08);
-            key_schedule[5] = KEYEXP128(key_schedule[4], 0x10);
-            key_schedule[6] = KEYEXP128(key_schedule[5], 0x20);
-            key_schedule[7] = KEYEXP128(key_schedule[6], 0x40);
-            key_schedule[8] = KEYEXP128(key_schedule[7], 0x80);
-            key_schedule[9] = KEYEXP128(key_schedule[8], 0x1B);
-            key_schedule[10] = KEYEXP128(key_schedule[9], 0x36);
+    
+#ifndef DISABLE_AESNI
+    if(!bCheckedAES)
+    {
+         uint32_t eax, ebx, ecx, edx;
+
+         eax = ebx = ecx = edx = 0;
+         __get_cpuid(1, &eax, &ebx, &ecx, &edx);
+         bHasAES=(ecx & bit_AES) > 0;
+         bCheckedAES=true;
+    }
+            
+    if(bHasAES)
+        return ni_aes_load_key(enc_key, keylen);
+#endif // DISABLE_AESNI
+
+    phex((char *)"secret",enc_key,keylen);
+    
+    switch(keylen){
+        case 32:
+            KeyExpansion(RoundKey256, enc_key, KEYNR256, KEYNK256);
+            phex((char *)"soft 256",RoundKey256,sizeof(RoundKey256));
             break;
-        }
-        case 32: {
-            /* 256 bit key setup */
-            key_schedule[0] = _mm_loadu_si128((const __m128i*) enc_key);
-            key_schedule[1] = _mm_loadu_si128((const __m128i*) (enc_key+16));
-            key_schedule[2] = KEYEXP256(key_schedule[0], key_schedule[1], 0x01);
-            key_schedule[3] = KEYEXP256_2(key_schedule[1], key_schedule[2]);
-            key_schedule[4] = KEYEXP256(key_schedule[2], key_schedule[3], 0x02);
-            key_schedule[5] = KEYEXP256_2(key_schedule[3], key_schedule[4]);
-            key_schedule[6] = KEYEXP256(key_schedule[4], key_schedule[5], 0x04);
-            key_schedule[7] = KEYEXP256_2(key_schedule[5], key_schedule[6]);
-            key_schedule[8] = KEYEXP256(key_schedule[6], key_schedule[7], 0x08);
-            key_schedule[9] = KEYEXP256_2(key_schedule[7], key_schedule[8]);
-            key_schedule[10] = KEYEXP256(key_schedule[8], key_schedule[9], 0x10);
-            key_schedule[11] = KEYEXP256_2(key_schedule[9], key_schedule[10]);
-            key_schedule[12] = KEYEXP256(key_schedule[10], key_schedule[11], 0x20);
-            key_schedule[13] = KEYEXP256_2(key_schedule[11], key_schedule[12]);
-            key_schedule[14] = KEYEXP256(key_schedule[12], key_schedule[13], 0x40);
+        case 16:
+            KeyExpansion(RoundKey128, enc_key, KEYNR128, KEYNK128);
+            phex((char *)"soft 128",RoundKey128,sizeof(RoundKey128));
             break;
-        }
     }
 }
 
-// Declares a global variable for efficiency.
-__m128i m_global;
+/*
+* XOR 128 bits
+*/
+static inline void xor128(const uint8_t *in1, const uint8_t *in2, uint8_t *out) {
+    for(int i=0;i<16;i++) {
+     out[i]=in1[i]^in2[i];
+    }
+}
 
 /*
  * Encrypts a plaintext using AES256.
  */
-static inline void aes256_enc(const uint8_t *plainText, uint8_t *cipherText) {
-    m_global = _mm_loadu_si128(reinterpret_cast<const __m128i *>(plainText));
+static inline void aes256_enc(const uint8_t *in, uint8_t *out) {
+    
+#ifndef DISABLE_AESNI
+    if(bHasAES)
+        return ni_aes256_enc(in, out);
+#endif // DISABLE_AESNI
+    
+    memcpy(out,in,16);
+    
+    state_t *state=(state_t*)out;
+    
+    uint8_t round = 0;
 
-    DO_ENC_BLOCK_256(m_global, key_schedule);
+    // Add the First round key to the state before starting the rounds.
+    AddRoundKey(0, state, RoundKey256);
 
-    _mm_storeu_si128(reinterpret_cast<__m128i *>(cipherText), m_global);
+    // There will be Nr rounds.
+    // The first Nr-1 rounds are identical.
+    // These Nr-1 rounds are executed in the loop below.
+    for (round = 1; round < ENCRYPTNR256; ++round)
+    {
+    SubBytes(state);
+    ShiftRows(state);
+    MixColumns(state);
+    AddRoundKey(round, state, RoundKey256);
+    }
+
+    // The last round is given below.
+    // The MixColumns function is not here in the last round.
+    SubBytes(state);
+    ShiftRows(state);
+    AddRoundKey(ENCRYPTNR256, state, RoundKey256);
 }
-
-/*
- * Encrypts a plaintext using AES128 with 2 rounds.
- */
-static inline void aes128_enc(const uint8_t *plainText, uint8_t *cipherText) {
-    m_global = _mm_loadu_si128(reinterpret_cast<const __m128i *>(plainText));
-
-    // Uses the 2 round encryption innstead of the full 10 round encryption
-    DO_ENC_BLOCK_2ROUND(m_global, key_schedule);
-
-    _mm_storeu_si128(reinterpret_cast<__m128i *>(cipherText), m_global);
-}
-
+    
 /*
  * Encrypts an integer using AES128 with 2 rounds.
  */
-static inline __m128i aes128_enc_int(__m128i plainText) {
-    // Uses the 2 round encryption innstead of the full 10 round encryption
-    DO_ENC_BLOCK_2ROUND(plainText, key_schedule);
-    return plainText;
+static inline void aes128_enc(uint8_t *in, uint8_t *out) {
+    
+#ifndef DISABLE_AESNI
+    if(bHasAES)
+        return ni_aes128_enc(in, out);
+#endif // DISABLE_AESNI
+    
+    memcpy(out,in,16);
+    
+    state_t *state=(state_t*)out;
+
+    uint8_t round = 0;
+
+    // Add the First round key to the state before starting the rounds.
+    AddRoundKey(0, state, RoundKey128);
+    
+    // There will be Nr rounds.
+    // The first Nr-1 rounds are identical.
+    // These Nr-1 rounds are executed in the loop below.
+    for (round = 1; round < ENCRYPTNR128; ++round)
+    {
+      SubBytes(state);
+      ShiftRows(state);
+      MixColumns(state);
+      AddRoundKey(round, state, RoundKey128);
+    }
 }
-
-__m128i m1;
-__m128i m2;
-__m128i m3;
-__m128i m4;
-
+    
 /*
  * Uses AES cache mode to map a 2 block ciphertext into 128 bit result.
  */
 static inline void aes128_2b(uint8_t *block1, uint8_t *block2, uint8_t *res) {
-    m1 = _mm_loadu_si128(reinterpret_cast<__m128i *>(block1));
-    m2 = _mm_loadu_si128(reinterpret_cast<__m128i *>(block2));
-    m3 = aes128_enc_int(m1);  // E(L)
-    m3 = aes128_enc_int(_mm_xor_si128(m3, m2));
-    _mm_storeu_si128(reinterpret_cast<__m128i *>(res), m3);
-}
+    
+#ifndef DISABLE_AESNI
+    if(bHasAES)
+        return ni_aes128_2b(block1, block2, res);
+#endif // DISABLE_AESNI
+   
+    uint8_t m1[16];
+    uint8_t m2[16];
+    uint8_t m3[16];
+    uint8_t intermediate[16];
 
+    memcpy(m1,block1,16);
+    memcpy(m2,block2,16);
+    
+    aes128_enc(m1,m3);
+    xor128(m3, m2, intermediate);
+    aes128_enc(intermediate,m3);
+    
+    memcpy(res,m3,16);
+}
+    
 /*
  * Uses AES cache mode to map a 3 block ciphertext into 128 bit result.
  */
 static inline void aes128_3b(uint8_t *block1, uint8_t* block2, uint8_t *block3, uint8_t* res) {
-    m1 = _mm_loadu_si128(reinterpret_cast<__m128i *>(block1));
-    m2 = _mm_loadu_si128(reinterpret_cast<__m128i *>(block2));
+    
+#ifndef DISABLE_AESNI
+    if(bHasAES)
+        return ni_aes128_3b(block1, block2, block3, res);
+#endif // DISABLE_AESNI
+ 
+    uint8_t m1[16];
+    uint8_t m2[16];
+    uint8_t m3[16];
 
-    m1 = aes128_enc_int(m1);  // E(La)
-    m2 = aes128_enc_int(m2);  // E(Ra)
+    memcpy(m1,block1,16);
+    memcpy(m2,block2,16);
 
-    m1 = _mm_xor_si128(m1, m2);
-    m2 = _mm_loadu_si128(reinterpret_cast<__m128i *>(block3));
-
-    m2 = aes128_enc_int(m2);
-    m1 = _mm_xor_si128(m1, m2);
-    m3 = aes128_enc_int(m1);
-    _mm_storeu_si128(reinterpret_cast<__m128i *>(res), m3);
+    aes128_enc(m1,m1); // E(La)
+    aes128_enc(m2,m2); // E(Ra)
+    
+    xor128(m1, m2, m1);
+    memcpy(m2,block3,16);
+    
+    aes128_enc(m2,m2);
+    xor128(m1, m2, m1);
+    aes128_enc(m1,m3);
+    memcpy(res,m3,16);
 }
 
 /*
  * Uses AES cache mode to map a 4 block ciphertext into 128 bit result.
  */
 static inline void aes128_4b(uint8_t *block1, uint8_t* block2, uint8_t *block3, uint8_t* block4, uint8_t* res) {
-    m1 = _mm_loadu_si128(reinterpret_cast<__m128i *>(block1));
-    m2 = _mm_loadu_si128(reinterpret_cast<__m128i *>(block3));
-    m3 = _mm_loadu_si128(reinterpret_cast<__m128i *>(block2));
-    m4 = _mm_loadu_si128(reinterpret_cast<__m128i *>(block4));
+   
+#ifndef DISABLE_AESNI
+    if(bHasAES)
+        return ni_aes128_4b(block1, block2, block3, block4, res);
+#endif // DISABLE_AESNI
 
-    m1 = aes128_enc_int(m1);    // E(La)
-    m1 = _mm_xor_si128(m1, m3);
-    m1 = aes128_enc_int(m1);    // E(E(La) ^ Lb)
-    m2 = aes128_enc_int(m2);    // E(Ra)
+    uint8_t m1[16];
+    uint8_t m2[16];
+    uint8_t m3[16];
+    uint8_t m4[16];
 
-    m1 = _mm_xor_si128(m1, m2);  // xor e(Ra)
-    m1 = _mm_xor_si128(m1, m4);  // xor Rb
+    memcpy(m1,block1,16);
+    memcpy(m2,block2,16);
+    memcpy(m3,block3,16);
+    memcpy(m4,block4,16);
 
-    m3 = aes128_enc_int(m1);
-    _mm_storeu_si128(reinterpret_cast<__m128i *>(res), m3);
+    aes128_enc(m1,m1); // E(La)
+    xor128(m1, m3, m1);
+    aes128_enc(m1,m1); // E(E(La) ^ Lb)
+    aes128_enc(m2,m2); // E(Ra)
+    
+    xor128(m1, m2, m1); // xor e(Ra)
+    xor128(m1, m4, m1); // xor Rb
+    
+    aes128_enc(m1,m3); // E(La)
+    memcpy(res,m3,16);
 }
 
-#endif  // SRC_CPP_AES_HPP_
